+++ include/save.h
@@ -1,12 +1,22 @@
 #ifndef __SAVE_H__
 #define __SAVE_H__
 
+typedef enum {
+    SAVETYPE_VANILLA,
+    SAVETYPE_MAX,
+} SaveType;
+
+typedef struct {
+    SaveType saveType;
+} ShipSaveData;
+
 typedef struct{
     u8 magic;
     u8 slotIndex;
     u8 data[0x70];
     u8 padding[0x2];
     u32 checksum;
+    ShipSaveData shipSaveData;
 }SaveData;
 
 typedef struct {
