+++ src/core2/fileselect.c
@@ -78,7 +78,6 @@ s32 gameFile_8033CFD4(s32 gamenum){
     s32 eeprom_error;
     SaveData *save_data;
 
-
     filenum = D_80383F04;
     next = gameFile_GameIdToFileIdMap[gamenum];
     gameFile_GameIdToFileIdMap[gamenum] = D_80383F04;
@@ -86,26 +85,28 @@ s32 gameFile_8033CFD4(s32 gamenum){
     save_data = gameFile_saveData + filenum;
     save_data->slotIndex = gamenum + 1;
     savedata_update_crc(save_data, sizeof(SaveData));
-    for(eeprom_error = 1; eeprom_error && i > 0; i--){//L8033D070
-        eeprom_error = savedata_8033CC98(filenum, save_data);
-        if(!eeprom_error){
-            __gameFile_8033CE14(gamenum);
+    CALL_CANCELLABLE_RETURN_EVENT(OnSaveFileSave, save_data, gamenum) {
+        for (eeprom_error = 1; eeprom_error && i > 0; i--) {//L8033D070
+            eeprom_error = savedata_8033CC98(filenum, save_data);
+            if (!eeprom_error) {
+                __gameFile_8033CE14(gamenum);
+            }
         }
-    }
-    if(!eeprom_error){
-        for(i = 3; i > 0; i--){//L8033D070
-            eeprom_error = savedata_8033CCD0(next);
-            if(!eeprom_error)
-                break;
+        if (!eeprom_error) {
+            for (i = 3; i > 0; i--) {//L8033D070
+                eeprom_error = savedata_8033CCD0(next);
+                if (!eeprom_error)
+                    break;
+            }
         }
+        if (eeprom_error) {
+            gameFile_GameIdToFileIdMap[gamenum] = next;
+        }
+        else {
+            D_80383F04 = next;
+        }
+        return eeprom_error;
     }
-    if(eeprom_error){
-        gameFile_GameIdToFileIdMap[gamenum] = next;
-    }
-    else{
-        D_80383F04 = next;
-    }
-    return eeprom_error;
 }
 
 void gameFile_clear(s32 gamenum){
@@ -113,7 +114,7 @@ void gameFile_clear(s32 gamenum){
     savedata_clear(&gameFile_saveData[filenum]);
 }
 
-extern void port_restoreFileEnhancementData(int eepromSlot);
+// extern void port_restoreFileEnhancementData(int eepromSlot);
 extern s32 D_80386068; // [port] lives backup — func_80347AA8 restores from this
 extern s32 D_80385F30[];
 extern void sns_set_item_state(s32 item, s32 set, bool value);
@@ -124,7 +125,7 @@ void gameFile_load(s32 gamenum){
     saveData_load(&gameFile_saveData[filenum]);
     // [port] Override lives backup and item array with persisted value.
     // func_80347AA8 reads D_80386068 to restore lives after map transitions.
-    port_restoreFileEnhancementData(filenum);
+    // port_restoreFileEnhancementData(filenum);
     D_80386068 = D_80385F30[ITEM_16_LIFE];
 
     // [port] Unlock Stop N' Swop items as a reward for 100% completion
