+++ src/core2/gameSelect.c
@@ -289,7 +289,7 @@ void func_802C4AF0(Actor * this){
     }
 
     for(i = 0; i < 3; i++){
-        gameFile_8033CFD4(i);
+        // gameFile_8033CFD4(i); Not needed with new Save System
     }
 
     if(D_8037DD28){
