+++ src/core2/savedata.c
@@ -328,17 +328,19 @@ void __savedata_save_abilities(u8 *savedata){ //savedata_save_abilities
 }
 
 s32 savedata_8033CA2C(s32 filenum, void *save_data_){
-    SaveData *save_data = (SaveData *)save_data_;
-    s32 sp1C;
+    CALL_CANCELLABLE_RETURN_EVENT(OnSaveFileLoad, filenum, save_data_, 0) {
+        SaveData* save_data = (SaveData*)save_data_;
+        s32 sp1C;
 
-    sp1C = eeprom_readBlocks(filenum, 0, save_data, 0xF);
-    if( sp1C 
-        || savedata_verify(0x78, save_data) 
-        || ((u8*)save_data)[baseOffset] != 0x11
-    ){
-        sp1C = 2;
+        sp1C = eeprom_readBlocks(filenum, 0, save_data, 0xF);
+        if (sp1C
+            || savedata_verify(0x78, save_data)
+            || ((u8*)save_data)[baseOffset] != 0x11
+            ) {
+            sp1C = 2;
+        }
+        return sp1C;
     }
-    return sp1C;
 }
 
 s32 savedata_8033CA9C(void *savedata_){
