+++ src/port/enhancements/events/hooks/EventSystem.h
@@ -73,14 +73,7 @@ typedef struct EventListener {
     eventType eventType##_ = { { false }, __VA_ARGS__ };                                    \
     EventSystem_CallEvent(eventType##ID, &eventType##_, __FILE__, __LINE__, FILE_AND_LINE); \
     if (eventType##_.event.cancelled) {                                                     \
-        return;                                                                             \
-    }
-
-#define CALL_CANCELLABLE_ACTORSPAWN_EVENT(eventType, ...)                                   \
-    eventType eventType##_ = { { false }, __VA_ARGS__ };                                    \
-    EventSystem_CallEvent(eventType##ID, &eventType##_, __FILE__, __LINE__, FILE_AND_LINE); \
-    if (eventType##_.event.cancelled) {                                                     \
-        return eventType##_.spawnedActor;                                                   \
+        return eventType##_.result;                                                                             \
     }
 
 #define REGISTER_VB_SHOULD(flag, body)                                            \
