+++ src/port/save/SaveConverter.h
@@ -1,0 +1,7 @@
+#ifndef SAVE_CONVERTER_H
+#define SAVE_CONVERTER_H
+#include "ship/Context.h"
+
+void SaveConverter_Init();
+
+#endif // SAVE_CONVERTER_H
